refactoring ImagePicker library structure
