+++ ImagePicker/src/main/java/com/jms/imagePicker/ui/picker/ImagePickerScaffold.kt
@@ -1,12 +1,9 @@
-package com.jms.imagePicker.ui
+package com.jms.imagePicker.ui.picker
 
+import android.content.Intent
 import android.net.Uri
-import android.os.Build
-import android.util.Log
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.animation.EnterTransition
-import androidx.compose.animation.ExitTransition
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.scrollBy
@@ -19,22 +16,16 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyGridState
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.rememberLazyGridState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.collectAsState
@@ -42,7 +33,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -56,12 +46,6 @@ import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import androidx.core.content.FileProvider
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavType
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
 import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
@@ -69,122 +53,25 @@ import androidx.paging.compose.itemKey
 import com.jms.imagePicker.Constants
 import com.jms.imagePicker.R
 import com.jms.imagePicker.component.ImageCell
-import com.jms.imagePicker.data.LocalGalleryDataSource
 import com.jms.imagePicker.extensions.photoGridDragHandler
-import com.jms.imagePicker.manager.API21MediaContentManager
-import com.jms.imagePicker.manager.API29MediaContentManager
-import com.jms.imagePicker.manager.FileManager
 import com.jms.imagePicker.model.Album
 import com.jms.imagePicker.model.MediaContent
-import com.jms.imagePicker.ui.preview.PreviewScreen
+import com.jms.imagePicker.ui.ImagePickerViewModel
 import com.jms.imagePicker.ui.scope.ImagePickerAlbumScope
-import com.jms.imagePicker.ui.scope.PreviewTopBarScope
+import com.jms.imagePicker.ui.scope.ImagePickerPreviewTopBarScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.drop
-import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 
-
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ImagePickerScreen(
-    state: ImagePickerState = rememberImagePickerState(),
-    albumTopBar: @Composable ImagePickerAlbumScope.() -> Unit = {},
-    previewTopBar: @Composable PreviewTopBarScope.() -> Unit = {},
-    content: @Composable BoxScope.(MediaContent) -> Unit
-) {
-    val context = LocalContext.current
-    val navController = rememberNavController()
-    Scaffold(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        NavHost(
-            modifier = Modifier.padding(it),
-            navController = navController,
-            startDestination = "route_image_list",
-            route = "graph_image_picker",
-            enterTransition = { EnterTransition.None },
-            exitTransition = { ExitTransition.None }
-        ) {
-            composable(
-                route = "route_image_list"
-            ) {
-                val parentEntry = remember(it) {
-                    navController.getBackStackEntry("graph_image_picker")
-                }
-
-                val viewModel: ImagePickerViewModel = viewModel(parentEntry) {
-                    ImagePickerViewModel(
-                        fileManager = FileManager(context = context.applicationContext),
-                        localGalleryDataSource = LocalGalleryDataSource(
-                            contentManager = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-                                API29MediaContentManager(context = context.applicationContext)
-                            } else
-                                API21MediaContentManager(context = context.applicationContext)
-                        )
-                    )
-                }
-
-                ImagePickerScaffold(
-                    viewModel = viewModel,
-                    state = state,
-                    albumTopBar = albumTopBar,
-                    content = content,
-                    previewTopBar = previewTopBar,
-                    onNavigateToPreview = {
-                        navController.navigate("route_preview?$it") {
-                            launchSingleTop = true
-
-                            popUpTo("route_image_list") {
-                                saveState = true
-                            }
-                        }
-                    }
-                )
-            }
-
-            composable(
-                route = "route_preview?{index}",
-                arguments = listOf(
-                    navArgument("index") {
-                        type = NavType.IntType
-                    }
-                )
-            ) {
-                val parentEntry = remember(it) {
-                    navController.getBackStackEntry("graph_image_picker")
-                }
-
-                val viewModel: ImagePickerViewModel = viewModel(parentEntry)
-                val initializeFirstVisibleItemIndex = it.arguments?.getInt("index") ?: 0
-
-                PreviewScreen(
-                    viewModel = viewModel,
-                    state = state,
-                    initializeFirstVisibleItemIndex = initializeFirstVisibleItemIndex,
-                    onBack = {
-                        navController.popBackStack(
-                            route = "route_image_list",
-                            inclusive = false
-                        )
-                    }
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun ImagePickerScaffold(
+internal fun ImagePickerScaffold(
     modifier: Modifier = Modifier,
     state: ImagePickerState = rememberImagePickerState(),
     viewModel: ImagePickerViewModel,
     onNavigateToPreview: (Int) -> Unit = {},
     albumTopBar: @Composable ImagePickerAlbumScope.() -> Unit = {},
-    previewTopBar: @Composable PreviewTopBarScope.() -> Unit = {},
+    previewTopBar: @Composable ImagePickerPreviewTopBarScope.() -> Unit = {},
     content: @Composable BoxScope.(MediaContent) -> Unit
 ) {
     val context = LocalContext.current
@@ -204,16 +91,11 @@ private fun ImagePickerScaffold(
 
     val cameraLaunch =
         rememberLauncherForActivityResult(contract = ActivityResultContracts.TakePicture()) {
-            if (it) {
-                viewModel.saveImageFile(
-                    max = state.max,
-                    autoSelectAfterCapture = state.autoSelectAfterCapture
-                )
-            }
+
         }
 
     val previewScopeImpl = remember(viewModel) {
-        object : PreviewTopBarScope {
+        object : ImagePickerPreviewTopBarScope {
             override val selectedMediaContents: List<MediaContent>
                 get() = selectedImages
 
@@ -246,9 +128,6 @@ private fun ImagePickerScaffold(
         ImagePickerContent(
             mediaContents = mediaContents,
             selectedUris = selectedUris,
-            onClick = {
-                viewModel.select(uri = it, max = state.max)
-            },
             onDragStart = {
                 viewModel.select(uri = it, max = state.max)
             },
@@ -271,9 +150,12 @@ private fun ImagePickerScaffold(
                     )
                 )
             },
-            onNavigateToPreview = { image ->
+            onClick = {
+                viewModel.select(uri = it.uri, max = state.max)
+            },
+            onNavigateToPreview = { mediaContent ->
                 val index =
-                    mediaContents.itemSnapshotList.indexOfFirst { it?.uri == image.uri }
+                    mediaContents.itemSnapshotList.indexOfFirst { it?.uri == mediaContent.uri }
                         .coerceAtLeast(0)
                 onNavigateToPreview(index)
             },
@@ -284,16 +166,16 @@ private fun ImagePickerScaffold(
 
 
 @Composable
-private fun ImagePickerContent(
+internal fun ImagePickerContent(
     modifier: Modifier = Modifier,
     mediaContents: LazyPagingItems<MediaContent>,
     selectedUris: List<Uri> = emptyList(),
-    onClick: (Uri) -> Unit,
     onPhoto: () -> Unit,
     onDragStart: (Uri) -> Unit,
     onDrag: (start: Int?, end: Int?, List<MediaContent>) -> Unit,
     onDragEnd: () -> Unit = {},
-    onNavigateToPreview: (MediaContent) -> Unit = { },
+    onClick: (MediaContent) -> Unit = {},
+    onNavigateToPreview: (MediaContent) -> Unit,
     content: @Composable BoxScope.(MediaContent) -> Unit
 ) {
     val gridState = rememberLazyGridState()
@@ -367,22 +249,19 @@ private fun ImagePickerContent(
             count = mediaContents.itemCount,
             key = mediaContents.itemKey { it.uri }
         ) {
-            mediaContents[it]?.let { content ->
+            mediaContents[it]?.let { mediaContent ->
                 Box(
                     modifier = Modifier
                         .clickable {
-                            onClick(content.uri)
+                            onClick(mediaContent)
                         }
                         .aspectRatio(1f)
                 ) {
                     ImageCell(
                         modifier = Modifier.fillMaxSize(),
                         cellDp = width.dp,
-                        mediaContent = content,
+                        mediaContent = mediaContent,
                     )
-
-                    if (content.selected)
-                        content(content)
 
                     Row(
                         modifier = Modifier.align(Alignment.BottomStart)
@@ -397,7 +276,7 @@ private fun ImagePickerContent(
                                         shape = RoundedCornerShape(5.dp)
                                     )
                                     .clickable {
-                                        onNavigateToPreview(content)
+                                        onNavigateToPreview(mediaContent)
                                     },
                                 painter = iconOfExpandContent,
                                 contentDescription = "expand_content",
@@ -406,6 +285,8 @@ private fun ImagePickerContent(
                             Spacer(modifier = Modifier.height(3.dp))
                         }
                     }
+
+                    content(mediaContent)
                 }
             }
         }
