+++ ImagePicker/src/main/java/com/jms/imagePicker/ui/scope/PreviewScope.kt
@@ -1,0 +1,9 @@
+package com.jms.imagePicker.ui.scope
+
+import androidx.compose.runtime.Stable
+
+
+@Stable
+interface PreviewScope {
+    fun onBack()
+}
